pyproject.toml
[project]
name = "dfs-analytics-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow==3.1.0",
    "duckdb>=1.4.0",
    "fsspec>=2025.9.0",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.9",
    "pandas>=2.3.3",
    "pyspark>=4.0.1",
    "python-dateutil>=2.8.2",
    "requests>=2.31.0",
    "s3fs>=2025.9.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
